Move split_command into a util module
