+++ cibuildwheel/util/cmd.py
@@ -4,7 +4,7 @@ import shutil
 import subprocess
 import sys
 import typing
-from collections.abc import Mapping
+from collections.abc import Iterator, Mapping
 from typing import Final, Literal
 
 from ..errors import FatalError
@@ -81,3 +81,18 @@ def shell(
     command = " ".join(commands)
     print(f"+ {command}")
     subprocess.run(command, env=env, cwd=cwd, shell=True, check=True)
+
+
+def split_command(lst: list[str]) -> Iterator[list[str]]:
+    """
+    Split a shell-style command, as returned by shlex.split, into a sequence
+    of commands, separated by '&&'.
+    """
+    items = list[str]()
+    for item in lst:
+        if item == "&&":
+            yield items
+            items = []
+        else:
+            items.append(item)
+    yield items
